Allow t/cmdline.t to test installed scripts
When running tests out of the source tree to validate installed
Tk-Pod, t/cmdline.t fails like this:

    $ xvfb-run -d perl -I. t/cmdline.t -d
    1..25
    /usr/bin/perl -Mblib=/usr/libexec/perl-Tk-Pod/t/../blib /usr/libexec/perl-Tk-Pod/t/../blib/script/tkpod -geometry +10+10 -tk
    Can't open perl script "/usr/libexec/perl-Tk-Pod/t/../blib/script/tkpod": No such file or directory
    not ok 1 - Trying tkpod with -tk
    #   Failed test 'Trying tkpod with -tk'
    #   at t/cmdline.t line 206.
    #          got: '512'
    #     expected: '0'

The cause is that the test hardcodes blib/script path the the
executed tkpod and tkmore.

This patch adds a fallback to $Config{installscript}, so that it falls
back to a system installation paths as loading module does.